requirements.txt
alabaster==0.7.12
aldryn-apphooks-config==0.5.2
aldryn-boilerplates==0.8.0
aldryn-categories==1.2.0
aldryn-common==1.0.5
aldryn-newsblog==2.2.1
aldryn-people==2.2.0
aldryn-translation-tools==0.3.0
anyjson==0.3.3
appdirs==1.4.3
argon2-cffi==19.1.0
arrow==0.15.2
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.7.0
backcall==0.1.0
backport-collections==0.1
binaryornot==0.4.4
black==19.3b0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cookiecutter==1.6.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.4
decorator==4.4.0
defusedxml==0.6.0
Django==2.1.8
django-admin-generator==2.1.0
django-admin-sortable2==0.7.3
django-admin-thumbnails==0.1.2
django-allauth==0.39.1
django-appconf==1.0.3
django-appdata==0.2.1
django-autoslug==1.9.6
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-daterangepicker==1.0.4
django-braces==1.13.0
django-classy-tags==0.9.0
django-cms==3.6.0
django-coverage-plugin==1.6.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-elasticsearch-dsl==6.4.2
django-environ==0.4.5
django-extensions==2.1.6
django-filer==1.5.0
django-formtools==2.1
django-haystack==2.8.1
django-js-asset==1.2.2
django-kindeditor==0.3.0
django-model-utils==3.1.2
django-mptt==0.10.0
django-oauth-toolkit==1.2.0
django-parler==2.0
django-phonenumber-field==1.3.0
django-polymorphic==2.0.3
django-redis==4.10.0
django-rest-framework-social-oauth2==1.1.0
django-sekizai==1.0.0
django-sortedm2m==2.0.0
django-taggit==1.1.0
django-tempus-dominus==5.1.2.8
django-tinymce==2.8.0
django-treebeard==4.3
django-utils2==2.5.4
djangocms-admin-style==1.4.0
djangocms-attributes-field==1.1.0
djangocms-column==1.9.0
djangocms-file==2.3.0
djangocms-googlemap==1.3.0
djangocms-link==2.5.0
djangocms-picture==2.3.0
djangocms-snippet==2.2.0
djangocms-style==2.2.0
djangocms-text-ckeditor==3.8.0
djangocms-video==2.1.1
djangorestframework==3.9.3
docutils==0.15.2
easy-thumbnails==2.6
elasticsearch==6.0.0
elasticsearch-dsl==6.4.0
entrypoints==0.3
factory-boy==2.11.1
Faker==0.9.1
flake8==3.7.5
future==0.17.1
html5lib==1.0.1
idna==2.8
imagesize==1.1.0
importlib-metadata==0.23
ipdb==0.12
ipython==7.8.0
ipython-genutils==0.2.0
isort==4.3.21
itypes==1.1.0
jedi==0.15.1
Jinja2==2.10.1
jinja2-time==0.2.0
lazy-object-proxy==1.4.2
lxml==4.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
mixer==6.1.3
model-mommy==1.6.0
more-itertools==7.2.0
mypy==0.701
mypy-extensions==0.4.1
mysqlclient==1.4.4
oauthlib==3.1.0
packaging==19.1
parso==0.5.1
pexpect==4.7.0
phonenumbers==8.10.18
phonenumberslite==8.10.18
pickleshare==0.7.5
Pillow==6.0.0
pluggy==0.13.0
polls==0.0.4
poyo==0.5.0
prompt-toolkit==2.0.9
psycopg2-binary==2.8.2
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.3.1
pylint-django==2.0.9
pylint-plugin-utils==0.5
pyparsing==2.4.2
pytest==4.4.1
pytest-cov==2.7.1
pytest-django==3.4.8
pytest-sugar==0.9.2
python-dateutil==2.8.0
python-slugify==3.0.2
python-utils==2.3.0
python3-openid==3.1.0
pytz==2019.1
redis==3.2.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
snowballstemmer==1.9.1
social-auth-app-django==3.1.0
social-auth-core==3.2.0
Sphinx==2.0.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
termcolor==1.1.0
text-unidecode==1.2
toml==0.10.0
traitlets==4.3.2
typed-ast==1.3.5
Unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
whichcraft==0.6.1
Whoosh==2.7.4
wrapt==1.11.2
YURL==1.0.0
zipp==0.6.0
